+++ isy-batchrahmen/src/main/java/de/bund/bva/isyfact/batchrahmen/batch/konstanten/KonfigurationSchluessel.java
@@ -17,70 +17,101 @@
 package de.bund.bva.isyfact.batchrahmen.batch.konstanten;
 
 /**
- * Enthält alle Schlüssel für Konfigurationsparameter.
- *
- *
+ * Contains all keys for configuration parameters.
  */
-public abstract class KonfigurationSchluessel {
+public final class KonfigurationSchluessel {
+
+    private KonfigurationSchluessel() {
+        // class with config keys should not be initialized
+    }
 
     /***************************************************************************
      * PROPERTY - PARAMETER *
      **************************************************************************/
 
-    /** Property-Name fuer die Batchrahmen-Bean. */
+    /**
+     * Property name for the Batchrahmen-bean.
+     */
     public static final String PROPERTY_BATCHRAHMEN_BEAN_NAME = "Batchrahmen.BeanName";
 
-    /** Property-Name fuer die Logback-Kontiguration. */
+    /**
+     * Property name for the logback configuration.
+     */
     public static final String PROPERTY_BATCHRAHMEN_LOGBACK_CONF = "Batchrahmen.LogbackConfigFile";
 
-    /** Property-Name fuer das Commit-Intervall. */
+    /**
+     * Property name for the commit interval.
+     */
     public static final String PROPERTY_BATCHRAHMEN_COMMIT_INTERVALL = "Batchrahmen.CommitIntervall";
 
-    /** Property-Name fuer das Commit-Intervall. */
+    /**
+     * Property name for the number of records to process.
+     */
     public static final String PROPERTY_BATCHRAHMEN_ZU_VERARBEITENDE_DATENSAETZE_ANZAHL =
-        "Batchrahmen.AnzahlZuVerarbeitendeDatensaetze";
+            "Batchrahmen.AnzahlZuVerarbeitendeDatensaetze";
 
     /**
-     * Property-Name fuer die Ergebnisdatei.
+     * Property name for the result file.
      */
     public static final String PROPERTY_BATCHRAHMEN_ERGEBNIS_DATEI = "Batchrahmen.Ergebnisdatei";
 
     /**
-     * Property-Name fuer das Clear-Intervall (clearen des Hibernate Sssion-Caches).
+     * Property name for the clear interval (clearing the Hibernate session cache).
      */
     public static final String PROPERTY_BATCHRAHMEN_CLEAR_INTERVALL = "Batchrahmen.ClearIntervall";
 
-    /** Property-Name fuer den Namen der Ausfuehrungs-Bean. */
+    /**
+     * Property name for the name of the Ausfuehrungs-bean.
+     */
     public static final String PROPERTY_AUSFUEHRUNGSBEAN = "AusfuehrungsBean";
 
-    /** Property-Name fuer die ID des Batches. */
+    /**
+     * Property name for the ID of the batch.
+     */
     public static final String PROPERTY_BATCH_ID = "BatchId";
 
-    /** Property-Name fuer den Namen des Batches. */
+    /**
+     * Property name for the name of the batch.
+     */
     public static final String PROPERTY_BATCH_NAME = "BatchName";
 
+    /**
+     * Property name for the oauth2ClientRegistrationId of the batch.
+     */
+    public static final String PROPERTY_BATCH_OAUTH2_CLIENT_REGISTRATION_ID = "batch.oauth2ClientRegistrationId";
+
     /***************************************************************************
-     * KOMMANDOZEILEN - PARAMETER *
+     * COMMAND LINE PARAMETERS *
      **************************************************************************/
 
-    /** Kommandozeilen-Parameter fuer die Property-Datei. */
+    /**
+     * Command line parameters for the property file.
+     */
     public static final String KOMMANDO_PARAM_PROP_DATEI = "cfg";
 
-    /** Kommandozeilen-Parameter fuer die Start-Art des Batches: Start. */
+    /**
+     * Command line parameter for the start type of the batch: Start.
+     */
     public static final String KOMMANDO_PARAM_START = "start";
 
-    /** Kommandozeilen-Parameter fuer die Start-Art des Batches: Restart. */
+    /**
+     * Command line parameter for the start type of the batch: Restart.
+     */
     public static final String KOMMANDO_PARAM_RESTART = "restart";
 
     /**
-     * Kommandozeilen-Parameter dafuer, dass auch bei Abbruch neu gestartet wird.
+     * Command line parameter for restarting even if aborted.
      */
     public static final String KOMMANDO_PARAM_IGNORIERE_RESTART = "ignoriereRestart";
 
-    /** Kommandozeilen-Parameter dafuer, dass auch bei Lauf neu gestartet wird. */
+    /**
+     * Command line parameter that states that restart is also done on run.
+     */
     public static final String KOMMANDO_PARAM_IGNORIERE_LAUF = "ignoriereLauf";
 
-    /** Kommandozeilen-Parameter fuer die maximale Laufzeit des Batches (optional). */
+    /**
+     * Command line parameter for the maximum runtime of the batch (optional).
+     */
     public static final String KOMMANDO_PARAM_LAUFZEIT = "laufzeit";
 
 }