+++ isy-batchrahmen/src/test/java/de/bund/bva/isyfact/batchrahmen/AnwendungTestConfig.java
@@ -1,37 +1,41 @@
 package de.bund.bva.isyfact.batchrahmen;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import javax.persistence.EntityManagerFactory;
 import javax.sql.DataSource;
 
-import de.bund.bva.isyfact.batchrahmen.sicherheit.AccessManagerStub;
-import de.bund.bva.isyfact.aufrufkontext.AufrufKontextFactory;
-import de.bund.bva.isyfact.aufrufkontext.AufrufKontextVerwalter;
-import de.bund.bva.isyfact.aufrufkontext.impl.AufrufKontextFactoryImpl;
-import de.bund.bva.isyfact.aufrufkontext.impl.AufrufKontextVerwalterImpl;
-import de.bund.bva.isyfact.sicherheit.Sicherheit;
-import de.bund.bva.isyfact.sicherheit.annotation.GesichertInterceptor;
-import de.bund.bva.isyfact.sicherheit.config.IsySicherheitConfigurationProperties;
-import de.bund.bva.isyfact.sicherheit.impl.SicherheitImpl;
 import org.h2.jdbcx.JdbcDataSource;
-import org.springframework.aop.Advisor;
-import org.springframework.aop.aspectj.AspectJExpressionPointcut;
-import org.springframework.aop.support.DefaultPointcutAdvisor;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.security.oauth2.client.OAuth2ClientProperties;
+import org.springframework.boot.autoconfigure.security.oauth2.client.OAuth2ClientPropertiesRegistrationAdapter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
+import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 import org.springframework.orm.jpa.vendor.Database;
 import org.springframework.orm.jpa.vendor.HibernateJpaDialect;
 import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.oauth2.client.registration.ClientRegistration;
+import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
+import org.springframework.security.oauth2.client.registration.InMemoryClientRegistrationRepository;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
+
+import de.bund.bva.isyfact.security.autoconfigure.IsyOAuth2ClientAutoConfiguration;
+import de.bund.bva.isyfact.security.autoconfigure.IsySecurityAutoConfiguration;
 
 @Configuration
 @EnableTransactionManagement
 @EnableAspectJAutoProxy
+@EnableAutoConfiguration
+@EnableWebSecurity
 public class AnwendungTestConfig {
 
 
@@ -69,41 +73,5 @@ public class AnwendungTestConfig {
         transactionManager.setEntityManagerFactory(emf);
         return transactionManager;
     }
-    @Bean
-    public AccessManagerStub accessManagerStub() {
-        return new AccessManagerStub();
-    }
 
-    @Bean
-    public AufrufKontextFactory aufrufKontextFactory() {
-        return new AufrufKontextFactoryImpl<>();
-    }
-
-    @Bean
-    public AufrufKontextVerwalter aufrufKontextVerwalter() {
-        return new AufrufKontextVerwalterImpl();
-    }
-
-    @Bean
-    public IsySicherheitConfigurationProperties isySicherheitConfigurationProperties() {
-        return new IsySicherheitConfigurationProperties();
-    }
-
-    @Bean
-    public Sicherheit sicherheit(AccessManagerStub accessManagerStub, AufrufKontextFactory aufrufKontextFactory, AufrufKontextVerwalter aufrufKontextVerwalter, IsySicherheitConfigurationProperties properties) {
-        return new SicherheitImpl("/resources/sicherheit/rollenrechte.xml", aufrufKontextVerwalter, aufrufKontextFactory, accessManagerStub, properties);
-    }
-
-    @Bean
-    public GesichertInterceptor gesichertInterceptor(Sicherheit sicherheit) {
-        GesichertInterceptor gesichertInterceptor = new GesichertInterceptor(sicherheit);
-        return gesichertInterceptor;
-    }
-
-    @Bean
-    public Advisor gesichertPointAdvisor(GesichertInterceptor gesichertInterceptor) {
-        AspectJExpressionPointcut pointcut = new AspectJExpressionPointcut();
-        pointcut.setExpression("@annotation(de.bund.bva.isyfact.sicherheit.annotation.Gesichert) || @within(de.bund.bva.isyfact.sicherheit.annotation.Gesichert)");
-        return new DefaultPointcutAdvisor(pointcut, gesichertInterceptor);
-    }
 }