+++ isy-batchrahmen/src/test/java/de/bund/bva/isyfact/batchrahmen/batch/GesicherterTestBatch2.java
@@ -17,87 +17,38 @@
 package de.bund.bva.isyfact.batchrahmen.batch;
 
 import java.util.Date;
+import java.util.Optional;
+
+import org.springframework.security.access.annotation.Secured;
 
 import de.bund.bva.isyfact.batchrahmen.batch.exception.BatchAusfuehrungsException;
 import de.bund.bva.isyfact.batchrahmen.batch.konfiguration.BatchKonfiguration;
 import de.bund.bva.isyfact.batchrahmen.batch.protokoll.BatchErgebnisProtokoll;
-import de.bund.bva.isyfact.batchrahmen.batch.rahmen.AuthenticationCredentials;
 import de.bund.bva.isyfact.batchrahmen.batch.rahmen.BatchStartTyp;
 import de.bund.bva.isyfact.batchrahmen.batch.rahmen.VerarbeitungsErgebnis;
-import de.bund.bva.isyfact.sicherheit.annotation.Gesichert;
 
 /**
- * Verhält sich wie der BasicTest-Batch - jedoch wird eine Autorisierung vor der Initialisierung durchgeführt.
+ * Behaves like the BasicTest batch - but authorization is performed before initialization.
  * <p>
- * Die Daten, die zur Autorisierung verwendet werden, können in der Konfiguration hinterlegt werden.
- * 
- * 
+ * The data used for authorization can be stored in the configuration.
  */
+@Secured("PRIV_Recht1")
 public class GesicherterTestBatch2 extends BasicTestBatch {
 
-    /** Konfigurationsschlüssel für den verwendeten Nutzer. */
-    public static final String BATCH_BENUTZER = "batch.benutzer";
-
-    /** Konfigurationsschlüssel für das verwendeten Psaswort. */
-    public static final String BATCH_PASSWORT = "batch.passwort";
-
-    /** Konfigurationsschlüssel für das verwendete Behördenkennzeichen. */
-    public static final String BATCH_BHKNZ = "batch.bhknz";
-
-
-
     /**
-     * Diese Methode wird gesichert. {@inheritDoc}
+     * This method is secured. {@inheritDoc}
      */
     public VerarbeitungsErgebnis verarbeiteSatz() throws BatchAusfuehrungsException {
         return super.verarbeiteSatz();
     }
 
     /**
-     * speichert die Referenz auf die Konfiguration. {@inheritDoc}
+     * saves the reference to the configuration. {@inheritDoc}
      */
-    @Gesichert("Recht")
     public int initialisieren(BatchKonfiguration konfiguration, long satzNummer, String dbKey,
-        BatchStartTyp startTyp, Date datumLetzterErfolg, BatchErgebnisProtokoll protokoll)
-        throws BatchAusfuehrungsException {
+                              BatchStartTyp startTyp, Date datumLetzterErfolg, BatchErgebnisProtokoll protokoll)
+            throws BatchAusfuehrungsException {
         return super
-            .initialisieren(konfiguration, satzNummer, dbKey, startTyp, datumLetzterErfolg, protokoll);
+                .initialisieren(konfiguration, satzNummer, dbKey, startTyp, datumLetzterErfolg, protokoll);
     }
-
-    /**
-     * Liefert das Feld {@link #batchBenutzerKennung} zurück.
-     * @return Wert von batchBenutzerKennung
-     */
-    protected String getBatchBenutzerKennung(BatchKonfiguration konfiguration) {
-        return konfiguration.getAsString(BATCH_BENUTZER);
-    }
-
-    /**
-     * Liefert das Feld {@link #batchBenutzerPasswort} zurück.
-     * @return Wert von batchBenutzerPasswort
-     */
-    protected String getBatchBenutzerPasswort(BatchKonfiguration konfiguration) {
-        return konfiguration.getAsString(BATCH_PASSWORT);
-    }
-
-    /**
-     * Liefert das Feld {@link #batchBenutzerBhknz} zurück.
-     * @return Wert von batchBenutzerBhknz
-     */
-    protected String getBatchBenutzerBhknz(BatchKonfiguration konfiguration) {
-        return konfiguration.getAsString(BATCH_BHKNZ);
-    }
-    
-    /**
-     * Dieser Batch verwendet keine Sicherung.
-     * {@inheritDoc}
-     */
-    public AuthenticationCredentials getAuthenticationCredentials(BatchKonfiguration konfiguration) {
-        AuthenticationCredentials authentifizierung = new AuthenticationCredentials();
-        authentifizierung.setBehoerdenkennzeichen(getBatchBenutzerBhknz(konfiguration));
-        authentifizierung.setBenutzerkennung(getBatchBenutzerKennung(konfiguration));
-        authentifizierung.setPasswort(getBatchBenutzerPasswort(konfiguration));
-        return authentifizierung;
-    }
-
 }