+++ isy-batchrahmen/src/test/java/de/bund/bva/isyfact/batchrahmen/batch/ReturnCodeTestBatch.java
@@ -28,52 +28,59 @@ import de.bund.bva.isyfact.batchrahmen.batch.protokoll.BatchErgebnisProtokoll;
 import de.bund.bva.isyfact.batchrahmen.batch.protokoll.MeldungTyp;
 import de.bund.bva.isyfact.batchrahmen.batch.protokoll.StatistikEintrag;
 import de.bund.bva.isyfact.batchrahmen.batch.protokoll.VerarbeitungsMeldung;
-import de.bund.bva.isyfact.batchrahmen.batch.rahmen.AuthenticationCredentials;
 import de.bund.bva.isyfact.batchrahmen.batch.rahmen.BatchAusfuehrungsBean;
 import de.bund.bva.isyfact.batchrahmen.batch.rahmen.BatchReturnCode;
 import de.bund.bva.isyfact.batchrahmen.batch.rahmen.BatchStartTyp;
 import de.bund.bva.isyfact.batchrahmen.batch.rahmen.VerarbeitungsErgebnis;
 
 /**
- * Dieser Batch setzt aber im Gegensatz zum Basic Test Batch den Return Code im Batchprotokoll auf einen Wert
- * ungleich 0 um. Mit diesem Batch soll getestet werden, dass der Batchrahmen einen gesetzten ReturnCode nicht
- * mehr überschreibt.
- *
+ * Unlike the Basic Test Batch, this batch converts the return code in the batch protocol to a value other than 0.
+ * This batch is used to test that the batch frame no longer overwrites a set return code.
  */
 public class ReturnCodeTestBatch implements BatchAusfuehrungsBean {
 
-    /** Die Anzahl der Sätze, die geschrieben werden sollen. */
+    /**
+     * The number of blocks to be written.
+     */
     public static final int MAX_SAETZE = 10000;
 
-    /** Das Ergebnisprotokoll */
+    /**
+     * The results log
+     */
     private BatchErgebnisProtokoll protokoll;
 
-    /** Der Logger. */
+    /**
+     * The logger.
+     */
     private static final IsyLogger LOG = IsyLoggerFactory.getLogger(ReturnCodeTestBatch.class);
 
-    /** StatistikEintraege fuer die Sollwertkontrolle */
+    /**
+     * {@link StatistikEintrag} for control.
+     */
     private StatistikEintrag statistik1;
 
+    /**
+     * {@link StatistikEintrag} for control.
+     */
     private StatistikEintrag statistik2;
 
-    /** Der interne Zaehler */
+    /**
+     * The internal counter
+     */
     private int count;
 
-    /** {@inheritDoc} */
     public void batchBeendet() {
         this.statistik2.setWert(999);
     }
 
-    /** {@inheritDoc} */
     public void checkpointGeschrieben(long satzNummer) throws BatchAusfuehrungsException {
         LOG.info(LogKategorie.JOURNAL, BatchRahmenEreignisSchluessel.EPLBAT00001,
-            "Checkpoint für Satz {} geschrieben.", satzNummer);
+                "Checkpoint für Satz {} geschrieben.", satzNummer);
     }
 
-    /** {@inheritDoc} */
     public int initialisieren(BatchKonfiguration konfiguration, long satzNummer, String dbKey,
-        BatchStartTyp startTyp, Date datumLetzterErfolg, BatchErgebnisProtokoll protokoll)
-        throws BatchAusfuehrungsException {
+                              BatchStartTyp startTyp, Date datumLetzterErfolg, BatchErgebnisProtokoll protokoll)
+            throws BatchAusfuehrungsException {
         this.count = MAX_SAETZE;
         this.protokoll = protokoll;
 
@@ -84,47 +91,32 @@ public class ReturnCodeTestBatch implements BatchAusfuehrungsBean {
         protokoll.registriereStatistikEintrag(this.statistik2);
 
         LOG.info(LogKategorie.JOURNAL, BatchRahmenEreignisSchluessel.EPLBAT00001, "Start-Typ ist: {}",
-            startTyp);
+                startTyp);
 
         return this.count;
     }
 
-    /** {@inheritDoc} */
     public VerarbeitungsErgebnis verarbeiteSatz() throws BatchAusfuehrungsException {
         this.count--;
         this.protokoll.ergaenzeMeldung(new VerarbeitungsMeldung("ID" + this.count, "REGNR" + this.count,
-            MeldungTyp.INFO, "Satz " + this.count + " verarbeitet. Teststring: <>\"üaößÜÄÖ€"));
+                MeldungTyp.INFO, "Satz " + this.count + " verarbeitet. Teststring: <>\"üaößÜÄÖ€"));
         this.statistik1.erhoeheWert();
         this.protokoll.setReturnCode(BatchReturnCode.FEHLER_AUSGEFUEHRT);
-        return new VerarbeitungsErgebnis("" + this.count, this.count == 0);
+        return new VerarbeitungsErgebnis(String.valueOf(this.count), this.count == 0);
     }
 
-    /** {@inheritDoc} */
     public void rollbackDurchgefuehrt() {
         LOG.error(BatchRahmenEreignisSchluessel.EPLBAT00001, "Rollback durchgeführt");
     }
 
-    /**
-     * Dieser Batch verwendet keine Sicherung. {@inheritDoc}
-     */
-    public AuthenticationCredentials getAuthenticationCredentials(BatchKonfiguration konfiguration) {
-        return null;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
     @Override
     public void vorCheckpointGeschrieben(long satzNummer) throws BatchAusfuehrungsException {
-        // leer
+        // blank
     }
 
-    /**
-     * {@inheritDoc}
-     */
     @Override
     public void vorRollbackDurchgefuehrt() {
-        // leer
+        // blank
     }
 
 }